requirements.txt
Flask
pinecone-client
langchain
langchain-google-genai
sentence-transformers